+++ cpp/brain/learner/RLPowerLearner.cpp
@@ -14,7 +14,7 @@ using namespace revolve::brain;
 
 RLPowerLearner::RLPowerLearner(std::string modelName,
                                Config brain,
-                               unsigned int n_weight_vectors) :
+                               size_t n_weight_vectors) :
         generation_counter_(0), n_weight_vectors_(n_weight_vectors), cycle_start_time_(-1), robot_name_(modelName)
 {
 
@@ -59,9 +59,9 @@ RLPowerLearner::generateInitPolicy()
   if (!current_policy_)
     current_policy_ = std::make_shared<Policy>(n_weight_vectors_);
 
-  for (unsigned int i = 0; i < n_weight_vectors_; i++) {
+  for (size_t i = 0; i < n_weight_vectors_; i++) {
     Spline spline(source_y_size_);
-    for (unsigned int j = 0; j < source_y_size_; j++) {
+    for (size_t j = 0; j < source_y_size_; j++) {
       spline[j] = dist(mt);
     }
     current_policy_->at(i) = spline;
@@ -85,7 +85,7 @@ RLPowerLearner::loadPolicy(std::string const policy_path)
   std::cout << "steps: " << policy_file[0]["steps"] << std::endl;
   std::cout << "velocity: " << policy_file[0]["population"][0]["velocity"] << std::endl;
 
-  unsigned int k = 0;
+  size_t k = 0;
   source_y_size_ = policy_file[0]["steps"].as<uint>();
   YAML::Node policy = policy_file[0]["population"][0]["policy"];
 
@@ -94,9 +94,9 @@ RLPowerLearner::loadPolicy(std::string const policy_path)
     return;
   }
 
-  for (unsigned int i = 0; i < n_weight_vectors_; i++) {
+  for (size_t i = 0; i < n_weight_vectors_; i++) {
     Spline spline(source_y_size_);
-    for (unsigned int j = 0; j < source_y_size_; j++) {
+    for (size_t j = 0; j < source_y_size_; j++) {
       spline[j] = policy[k++].as<double>();
     }
     current_policy_->at(i) = spline;
@@ -110,7 +110,7 @@ RLPowerLearner::reportFitness(std::string id,
 {
   // Insert ranked policy in list
   PolicyPtr policy_copy = std::make_shared<Policy>(n_weight_vectors_);
-  for (unsigned int i = 0; i < n_weight_vectors_; i++) {
+  for (size_t i = 0; i < n_weight_vectors_; i++) {
     Spline &spline = current_policy_->at(i);
     policy_copy->at(i) = Spline(spline.begin(),
                                 spline.end());
@@ -174,8 +174,8 @@ RLPowerLearner::reportFitness(std::string id,
   /// For algorithms B and D, is used two parent crossover with binary tournament selection
   if (ranked_policies_.size() < max_ranked_policies_) {
     // Generate random policy if number of stored policies is less then 'max_ranked_policies_'
-    for (unsigned int i = 0; i < n_weight_vectors_; i++) {
-      for (unsigned int j = 0; j < source_y_size_; j++) {
+    for (size_t i = 0; i < n_weight_vectors_; i++) {
+      for (size_t j = 0; j < source_y_size_; j++) {
         (*current_policy_)[i][j] = dist(mt);
       }
     }
@@ -200,9 +200,9 @@ RLPowerLearner::reportFitness(std::string id,
       total_fitness = fitness1 + fitness2;
 
       // For each spline
-      for (unsigned int i = 0; i < n_weight_vectors_; i++) {
+      for (size_t i = 0; i < n_weight_vectors_; i++) {
         // And for each control point
-        for (unsigned int j = 0; j < source_y_size_; j++) {
+        for (size_t j = 0; j < source_y_size_; j++) {
           // Apply modifier
           double spline_point = 0;
           spline_point += ((policy1->at(i)[j] - (*current_policy_)[i][j])) * (fitness1 / total_fitness);
@@ -227,9 +227,9 @@ RLPowerLearner::reportFitness(std::string id,
 
       // For each spline
       // TODO: Verify that this should is correct formula
-      for (unsigned int i = 0; i < n_weight_vectors_; i++) {
+      for (size_t i = 0; i < n_weight_vectors_; i++) {
         // And for each control point
-        for (unsigned int j = 0; j < source_y_size_; j++) {
+        for (size_t j = 0; j < source_y_size_; j++) {
 
           // Apply modifier
           double spline_point = 0;
@@ -263,10 +263,10 @@ void
 RLPowerLearner::interpolateCubic(Policy *const source_y,
                                  Policy *destination_y)
 {
-  const unsigned int source_y_size_ = (*source_y)[0].size();
-  const unsigned int destination_y_size = (*destination_y)[0].size();
+  const size_t source_y_size_ = (*source_y)[0].size();
+  const size_t destination_y_size = (*destination_y)[0].size();
 
-  const unsigned int N = source_y_size_ + 1;
+  const size_t N = source_y_size_ + 1;
   double *x = new double[N];
   double *y = new double[N];
   double *x_new = new double[destination_y_size];
@@ -278,23 +278,23 @@ RLPowerLearner::interpolateCubic(Policy *const source_y,
 
   // init x
   double step_size = CYCLE_LENGTH / source_y_size_;
-  for (unsigned int i = 0; i < N; i++) {
+  for (size_t i = 0; i < N; i++) {
     x[i] = step_size * i;
   }
 
   // init x_new
   step_size = CYCLE_LENGTH / destination_y_size;
-  for (unsigned int i = 0; i < destination_y_size; i++) {
+  for (size_t i = 0; i < destination_y_size; i++) {
     x_new[i] = step_size * i;
   }
 
-  for (unsigned int j = 0; j < n_weight_vectors_; j++) {
+  for (size_t j = 0; j < n_weight_vectors_; j++) {
     Spline &source_y_line = source_y->at(j);
     Spline &destination_y_line = destination_y->at(j);
 
     // init y
     // TODO use memcpy
-    for (unsigned int i = 0; i < source_y_size_; i++) {
+    for (size_t i = 0; i < source_y_size_; i++) {
       y[i] = source_y_line[i];
     }
 
@@ -306,7 +306,7 @@ RLPowerLearner::interpolateCubic(Policy *const source_y,
                     y,
                     N);
 
-    for (unsigned int i = 0; i < destination_y_size; i++) {
+    for (size_t i = 0; i < destination_y_size; i++) {
       destination_y_line[i] = gsl_spline_eval(spline,
                                               x_new[i],
                                               acc);
@@ -332,7 +332,7 @@ RLPowerLearner::increaseSplinePoints()
 
   // Copy current policy for resizing
   Policy policy_copy(current_policy_->size());
-  for (unsigned int i = 0; i < n_weight_vectors_; i++) {
+  for (size_t i = 0; i < n_weight_vectors_; i++) {
     Spline &spline = current_policy_->at(i);
     policy_copy[i] = Spline(spline.begin(),
                             spline.end());
@@ -347,7 +347,7 @@ RLPowerLearner::increaseSplinePoints()
   for (auto &it : ranked_policies_) {
     PolicyPtr policy = it.second;
 
-    for (unsigned int j = 0; j < n_weight_vectors_; j++) {
+    for (size_t j = 0; j < n_weight_vectors_; j++) {
       Spline &spline = policy->at(j);
       policy_copy[j] = Spline(spline.begin(),
                               spline.end());
@@ -363,11 +363,11 @@ RLPowerLearner::binarySelection()
 {
   std::random_device rd;
   std::mt19937 umt(rd());
-  std::uniform_int_distribution<unsigned int> udist(0,
+  std::uniform_int_distribution<size_t> udist(0,
                                                     max_ranked_policies_ - 1);
 
   // Select two different numbers from uniform distribution U(0, max_ranked_policies_ - 1)
-  unsigned int pindex1, pindex2;
+  size_t pindex1, pindex2;
   pindex1 = udist(umt);
   do {
     pindex2 = udist(umt);
@@ -418,9 +418,9 @@ RLPowerLearner::writeElite()
     PolicyPtr policy = it.second;
     outputFile << "   - velocity: " << fitness << std::endl;
     outputFile << "     policy:" << std::endl;
-    for (unsigned int i = 0; i < policy->size(); i++) {
+    for (size_t i = 0; i < policy->size(); i++) {
       Spline &spline = policy->at(i);
-      for (unsigned int j = 0; j < spline.size(); j++) {
+      for (size_t j = 0; j < spline.size(); j++) {
         outputFile << "      - " << spline.at(j) << std::endl;
       }
     }
@@ -428,11 +428,11 @@ RLPowerLearner::writeElite()
   outputFile.close();
 }
 
-const unsigned int RLPowerLearner::MAX_EVALUATIONS = 1000; // max number of evaluations
-const unsigned int RLPowerLearner::MAX_RANKED_POLICIES = 10; // max length of policies vector
-const unsigned int RLPowerLearner::INTERPOLATION_CACHE_SIZE = 100; // number of data points for the interpolation cache
-const unsigned int RLPowerLearner::INITIAL_SPLINE_SIZE = 3; // number of initially sampled spline points
-const unsigned int RLPowerLearner::UPDATE_STEP = 100; // after # generations, it increases the number of spline points
+const size_t RLPowerLearner::MAX_EVALUATIONS = 1000; // max number of evaluations
+const size_t RLPowerLearner::MAX_RANKED_POLICIES = 10; // max length of policies vector
+const size_t RLPowerLearner::INTERPOLATION_CACHE_SIZE = 100; // number of data points for the interpolation cache
+const size_t RLPowerLearner::INITIAL_SPLINE_SIZE = 3; // number of initially sampled spline points
+const size_t RLPowerLearner::UPDATE_STEP = 100; // after # generations, it increases the number of spline points
 const double RLPowerLearner::EVALUATION_RATE = 30.0; // evaluation time for each policy
 const double RLPowerLearner::SIGMA_START_VALUE = 0.8; // starting value for sigma
 const double RLPowerLearner::SIGMA_TAU_CORRECTION = 0.2;
